Refactored into story
