+++ packages/shared/src/types/Story.ts
@@ -1,0 +1,14 @@
+export interface Page {
+  text: string;
+  illustrationBase64: string;
+}
+
+export interface Story {
+  id: string;
+  title: string;
+  summary?: string;
+  coverImageBase64: string;
+  pages: Page[];
+  createdAt?: Date;
+  isPreinstalled?: boolean;
+}
